Crate author_inherent

Pallet that allows block authors to include their identity in a block via an inherent. Currently the author does not prove their identity, just states it. So it should not be used, for things like equivocation slashing that require authenticated authorship information.

Structs

InherentDataProvider
The thing that the outer node will use to actually inject the inherent data

Module

Enums

Call
Dispatchable calls.

Error

InherentError

Constants

ENGINE_ID

INHERENT_IDENTIFIER

Traits

CanAuthor
Permissions for what block author can be set in this pallet

Config

EventHandler
The given account ID is the author of the current block.

_
Provides version information about a pallet.

Type Definitions

InherentType
The type of data that the inherent will contain. Just a byte array. It will be decoded to an actual account id later.