Trait author_inherent::Config

pub trait Config: System {
    type EventHandler: EventHandler<Self::AccountId>;
    type PreliminaryCanAuthor: CanAuthor<Self::AccountId>;
    type FinalCanAuthor: CanAuthor<Self::AccountId>;
}

Associated Types

type EventHandler: EventHandler<Self::AccountId>

Other pallets that want to be informed about block authorship

type PreliminaryCanAuthor: CanAuthor<Self::AccountId>

A preliminary means of checking the validity of this author. This check is run before block execution begins when data from previous inherent is unavailable. This is meant to quickly invalidate blocks from obviously-invalid authors, although it need not rule out all invlaid authors. The final check will be made when executing the inherent.

type FinalCanAuthor: CanAuthor<Self::AccountId>

The final word on whether the reported author can author at this height. This will be used when executing the inherent. This check is often stricter than the Preliminary check, because it can use more data. If the pallet that implements this trait depends on an inherent, that inherent must be included before this one.

Loading content...